move tuidelete to its own package
